backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-multipart==0.0.17
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
pydantic[email]==2.10.3
pydantic-settings==2.6.1
statsmodels==0.14.4
scikit-learn==1.5.2
numpy==2.1.3
Pillow==11.0.0
pytesseract==0.3.13
